frontend/pyproject.toml
[project]
name = "PolyRag-Frontend"
version = "0.1.0"
description = "Frontend application for PolyRAG - Streamlit-based chat interface"
readme = "README.md"
authors = [{ name = "Joshua Carroll", email = "carroll.joshk@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

requires-python = ">=3.12"

dependencies = [
    "httpx==0.28.1",
    "pydantic~=2.10.1",
    "python-dotenv==1.1.0",
    "plotly>=6.0.1",
    "streamlit~=1.47.0",
    "streamlit-pdf-viewer~=0.0.26",
    "bcrypt>=4.3.0",
    "passlib>=1.7.4",
    "streamlit-cookies-controller==0.0.4",

]

[dependency-groups]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-env",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "U"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
plugins = "pydantic.mypy"
exclude = "src/streamlit-app.py"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
